done refactoring not tested exhaustively
